refactor: 패키지명 변경
step_2, step_3 -> calculator, racinggame
