+++ src/main/java/calculator/StringUtils.java
@@ -1,4 +1,4 @@
-package step_2;
+package calculator;
 
 public class StringUtils {
     private StringUtils() {}
