+++ src/main/java/racinggame/RaceConfigurationScanner.java
@@ -1,6 +1,6 @@
-package step_3;
+package racinggame;
 
-import step_3.dto.RaceConfigurationDTO;
+import racinggame.dto.RaceConfigurationDTO;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
