+++ src/main/java/racinggame/RaceCourse.java
@@ -1,6 +1,6 @@
-package step_3;
+package racinggame;
 
-import step_3.dto.RaceConfigurationDTO;
+import racinggame.dto.RaceConfigurationDTO;
 
 import java.util.List;
