+++ src/main/java/racinggame/RaceResultView.java
@@ -1,4 +1,4 @@
-package step_3;
+package racinggame;
 
 import java.util.List;
