+++ src/test/java/racinggame/CarTest.java
@@ -1,4 +1,4 @@
-package step_3;
+package racinggame;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
